베타 1.0.0.5
== 능력 ==
코드 정리

== 시너지 ==
코드 정리

== 그외 ==
AbilityWar 2.1.7.8 지원
Test 어노테이션 삭제
자기장 능력자 믹스 전쟁 추가

+++ src/cokes86/addon/ability/synergy/RoyalStraightFlush.java
@@ -47,9 +47,9 @@ public class RoyalStraightFlush extends Synergy implements ActiveHandler {
 				int r = range.getValue();
 				int a = new Random().nextInt(20), b = new Random().nextInt(20);
 				getPlayer().sendMessage("숫자를 뽑습니다 : "+a+", "+b);
-				getPlayer().sendMessage("총 "+(double)(a+b)/2+"대미지를 상대방에게 줍니다.");
+				getPlayer().sendMessage("총 "+(a+b)/2.0+"대미지를 상대방에게 줍니다.");
 				for (Player p : LocationUtil.getNearbyPlayers(getPlayer(), r, r)) {
-					p.damage((a+b)/2);
+					p.damage((a+b)/2.0);
 					SoundLib.ENTITY_FIREWORK_ROCKET_TWINKLE_FAR.playSound(p);
 				}
 				cooldown.start();
